requirements.txt
streamlit
tensorflow
numpy
pillowmediapipe